+++ JavaSource/org/unitime/timetable/form/TimePatternEditForm.java
@@ -268,7 +268,7 @@ public class TimePatternEditForm extends ActionForm {
 	
 	public TimePattern saveOrUpdate(SessionContext context, org.hibernate.Session hibSession) throws Exception {
 		TimePattern tp = null;
-		if (getUniqueId().intValue()>=0)
+		if (getUniqueId()>=0)
 			tp = (new TimePatternDAO()).get(getUniqueId());
 		if (tp==null) {
 			tp = create(context, hibSession);
@@ -295,7 +295,7 @@ public class TimePatternEditForm extends ActionForm {
 	}
 	
 	public void delete(SessionContext context, org.hibernate.Session hibSession) throws Exception {
-		if (getUniqueId().intValue()<0)
+		if (getUniqueId()<0)
 			return;
 		if (!getEditable())
 			return;
